% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{az_branches_list}
\alias{az_branches_list}
\title{List Branches in a Repository}
\usage{
az_branches_list(repository_id, project = NULL, client = NULL)
}
\arguments{
\item{repository_id}{Repository name or ID.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A \code{tibble} of branch names and commit object IDs.
}
\description{
List Branches in a Repository
}
