% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{az_browse}
\alias{az_browse}
\alias{az_browse,azureops::az_work_item-method}
\alias{az_browse,azureops::az_repo-method}
\alias{az_browse,azureops::az_pull_request-method}
\alias{az_browse,azureops::az_pipeline-method}
\alias{az_browse,azureops::az_pipeline_run-method}
\alias{az_browse,azureops::az_test_plan-method}
\alias{az_browse,azureops::az_test_run-method}
\title{Browse an Azure DevOps Resource in Web UI}
\usage{
az_browse(x, browser = interactive(), ...)

## S7 method for class <azureops::az_work_item>
az_browse(x, browser = interactive(), ..., call = rlang::caller_env())

## S7 method for class <azureops::az_repo>
az_browse(x, browser = interactive(), ..., call = rlang::caller_env())

## S7 method for class <azureops::az_pull_request>
az_browse(x, browser = interactive(), ..., call = rlang::caller_env())

## S7 method for class <azureops::az_pipeline>
az_browse(x, browser = interactive(), ..., call = rlang::caller_env())

## S7 method for class <azureops::az_pipeline_run>
az_browse(x, browser = interactive(), ..., call = rlang::caller_env())

## S7 method for class <azureops::az_test_plan>
az_browse(x, browser = interactive(), ..., call = rlang::caller_env())

## S7 method for class <azureops::az_test_run>
az_browse(x, browser = interactive(), ..., call = rlang::caller_env())
}
\arguments{
\item{x}{An Azure DevOps S7 object with a web URL.}

\item{browser}{Logical; whether to open in browser (defaults to \code{interactive()}).}

\item{...}{Additional arguments passed to methods.}
}
\value{
The URL string (invisibly if browser is opened).
}
\description{
Opens the corresponding Azure DevOps web page in the default web browser.
}
