% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{az_cancel}
\alias{az_cancel}
\title{Cancel an Azure DevOps Resource Operation}
\usage{
az_cancel(x, client = NULL, ...)
}
\arguments{
\item{x}{An Azure DevOps S7 object.}

\item{client}{An \code{az_client} object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Updated S7 object.
}
\description{
Cancels an in-progress pipeline run or closes/abandons a pull request.
}
