% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{az_commit_get}
\alias{az_commit_get}
\title{Get a Single Commit Details}
\usage{
az_commit_get(repository_id, commit_id, project = NULL, client = NULL)
}
\arguments{
\item{repository_id}{Repository name or ID.}

\item{commit_id}{Full SHA or prefix of commit.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A list representing commit details.
}
\description{
Get a Single Commit Details
}
