% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{az_pipeline}
\alias{az_pipeline}
\title{Azure DevOps Pipeline Definition Class}
\usage{
az_pipeline(
  id = integer(0),
  name = character(0),
  folder = "\\\\",
  revision = 1L,
  web_url = ""
)
}
\description{
Represents a build/release pipeline definition in Azure Pipelines.
}
