% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{az_pipeline_run_get}
\alias{az_pipeline_run_get}
\title{Get a Single Pipeline Run}
\usage{
az_pipeline_run_get(
  pipeline_id,
  run_id,
  project = NULL,
  client = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{pipeline_id}{Integer ID of the pipeline.}

\item{run_id}{Integer ID of the run.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}

\item{call}{Caller environment for error attribution.}
}
\value{
An S7 \code{az_pipeline_run} object.
}
\description{
Get a Single Pipeline Run
}
