% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{az_pipeline_run_log_get}
\alias{az_pipeline_run_log_get}
\title{Get Log Content by Log ID}
\usage{
az_pipeline_run_log_get(
  pipeline_id,
  run_id,
  log_id,
  project = NULL,
  client = NULL
)
}
\arguments{
\item{pipeline_id}{Integer ID of the pipeline.}

\item{run_id}{Integer ID of the run.}

\item{log_id}{Integer ID of the log chunk.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A character vector of log lines.
}
\description{
Get Log Content by Log ID
}
