% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{az_pipeline_run_logs_list}
\alias{az_pipeline_run_logs_list}
\title{List Execution Logs for a Pipeline Run}
\usage{
az_pipeline_run_logs_list(pipeline_id, run_id, project = NULL, client = NULL)
}
\arguments{
\item{pipeline_id}{Integer ID of the pipeline.}

\item{run_id}{Integer ID of the run.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A \code{tibble} of log entries.
}
\description{
List Execution Logs for a Pipeline Run
}
