% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{az_pipeline_runs_list}
\alias{az_pipeline_runs_list}
\title{List Pipeline Runs}
\usage{
az_pipeline_runs_list(pipeline_id = NULL, project = NULL, client = NULL)
}
\arguments{
\item{pipeline_id}{Optional integer ID of the pipeline.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A \code{tibble} of pipeline runs.
}
\description{
List Pipeline Runs
}
