% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{az_pipelines_list}
\alias{az_pipelines_list}
\title{List Azure Pipelines}
\usage{
az_pipelines_list(project = NULL, client = NULL)
}
\arguments{
\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A \code{tibble} of pipelines.
}
\description{
List Azure Pipelines
}
\examples{
\dontrun{
az_pipelines_list()
}
}
