% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{az_project}
\alias{az_project}
\title{Retrieve Azure DevOps Default Project}
\usage{
az_project(project = NULL)
}
\arguments{
\item{project}{Optional project string.}
}
\value{
A character string (empty string if not specified).
}
\description{
Retrieves the default project name from the provided argument or
\code{AZURE_DEVOPS_PROJECT} / \code{AZURE_PROJECT} environment variables.
}
