% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{az_pull_request}
\alias{az_pull_request}
\title{Azure DevOps Pull Request Class}
\usage{
az_pull_request(
  id = integer(0),
  title = character(0),
  status = character(0),
  source_branch = "",
  target_branch = "",
  created_by = "",
  web_url = "",
  repository = ""
)
}
\description{
Represents a pull request in Azure Repos.
}
