% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{az_pull_request_reviewers_get}
\alias{az_pull_request_reviewers_get}
\title{List Pull Request Reviewers}
\usage{
az_pull_request_reviewers_get(
  repository_id,
  pull_request_id,
  project = NULL,
  client = NULL
)
}
\arguments{
\item{repository_id}{Repository name or ID.}

\item{pull_request_id}{Integer ID of the pull request.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A \code{tibble} of reviewers and vote statuses.
}
\description{
List Pull Request Reviewers
}
