% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{az_repos_list}
\alias{az_repos_list}
\title{List Git Repositories}
\usage{
az_repos_list(project = NULL, client = NULL)
}
\arguments{
\item{project}{Project name or ID (optional).}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A \code{tibble} of repositories.
}
\description{
List Git Repositories
}
\examples{
\dontrun{
az_repos_list()
}
}
