% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boards.R
\name{az_sprint_capacity_get}
\alias{az_sprint_capacity_get}
\title{Get Sprint Capacity}
\usage{
az_sprint_capacity_get(
  iteration_id,
  team = NULL,
  project = NULL,
  client = NULL
)
}
\arguments{
\item{iteration_id}{UUID or ID of the iteration.}

\item{team}{Team name or ID.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A \code{tibble} of team member capacities.
}
\description{
Retrieves team and member capacity metrics for a given iteration/sprint.
}
