% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{az_team_members_list}
\alias{az_team_members_list}
\title{List Team Members}
\usage{
az_team_members_list(project, team_id, client = NULL)
}
\arguments{
\item{project}{Project name or ID.}

\item{team_id}{Team name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A \code{tibble} of team members.
}
\description{
List Team Members
}
