% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{az_teams_list}
\alias{az_teams_list}
\title{List Teams in a Project or Organization}
\usage{
az_teams_list(project = NULL, client = NULL)
}
\arguments{
\item{project}{Project name or ID (optional).}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A \code{tibble} of teams.
}
\description{
List Teams in a Project or Organization
}
