% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_plans.R
\name{az_test_cases_list}
\alias{az_test_cases_list}
\title{List Test Cases in a Test Suite}
\usage{
az_test_cases_list(plan_id, suite_id, project = NULL, client = NULL)
}
\arguments{
\item{plan_id}{Integer ID of the test plan.}

\item{suite_id}{Integer ID of the test suite.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A \code{tibble} of test cases.
}
\description{
List Test Cases in a Test Suite
}
