% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_plans.R
\name{az_test_plans_list}
\alias{az_test_plans_list}
\title{List Test Plans}
\usage{
az_test_plans_list(project = NULL, client = NULL)
}
\arguments{
\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A \code{tibble} of test plans.
}
\description{
Retrieves manual and automated test plans in a project.
}
\examples{
\dontrun{
az_test_plans_list()
}
}
