% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_plans.R
\name{az_test_suites_list}
\alias{az_test_suites_list}
\title{List Test Suites for a Test Plan}
\usage{
az_test_suites_list(plan_id, project = NULL, client = NULL)
}
\arguments{
\item{plan_id}{Integer ID of the test plan.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
A \code{tibble} of test suites.
}
\description{
List Test Suites for a Test Plan
}
