% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azureops-package.R
\docType{package}
\name{azureops-package}
\alias{azureops}
\alias{azureops-package}
\title{azureops: Interface with the Azure DevOps REST API}
\description{
Programmatically interface with the Azure DevOps REST API for querying resources and triggering automated workflows across Azure Boards, Repos, Pipelines, Test Plans, and Administration using 'httr2' and 'S7'.
}
\author{
\strong{Maintainer}: Matt \email{matt@example.com}

Authors:
\itemize{
  \item Matt \email{matt@example.com}
}

}
\keyword{internal}
