% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{az_org}
\alias{az_org}
\title{Retrieve Azure DevOps Default Organization}
\usage{
az_org(organization = NULL, call = rlang::caller_env())
}
\arguments{
\item{organization}{Optional organization string.}
}
\value{
An organization string.
}
\description{
Retrieves the organization name from the provided argument, or checks
the \code{AZURE_DEVOPS_ORG} and \code{AZURE_ORG} environment variables.
}
