% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{az_pat}
\alias{az_pat}
\title{Retrieve or Validate Azure DevOps Personal Access Token (PAT)}
\usage{
az_pat(pat = NULL)
}
\arguments{
\item{pat}{Optional PAT string. If \code{NULL}, looks in environment variables.}
}
\value{
A non-empty PAT string.
}
\description{
Retrieves the PAT from the provided argument, or checks the
\code{AZURE_DEVOPS_PAT} and \code{AZURE_PAT} environment variables.
}
\examples{
\dontrun{
az_pat()
}
}
