% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{az_pipeline}
\alias{az_pipeline}
\title{Azure DevOps Pipeline Definition Class}
\usage{
az_pipeline(
  id = integer(0),
  name = character(0),
  folder = "\\\\",
  revision = 1L,
  web_url = ""
)
}
\arguments{
\item{id}{Integer ID of the pipeline (\code{integer}).}

\item{name}{Name of the pipeline (\code{character}).}

\item{folder}{Folder path where the pipeline definition resides (\code{character}).}

\item{revision}{Integer revision number of the pipeline definition (\code{integer}).}

\item{web_url}{Web browser URL for the pipeline in Azure Pipelines (\code{character}).}
}
\value{
An S7 \code{az_pipeline} object.
}
\description{
An S7 class representing a build or release pipeline definition in Azure Pipelines.
Corresponds to the Azure DevOps Pipeline REST API resource (\verb{/_apis/pipelines}).
}
