% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{az_pipeline_get}
\alias{az_pipeline_get}
\title{Get a Single Pipeline Definition}
\usage{
az_pipeline_get(
  pipeline_id,
  project = NULL,
  client = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{pipeline_id}{Integer ID of the pipeline.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
An S7 \code{az_pipeline} object.
}
\description{
Get a Single Pipeline Definition
}
