% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{az_pipeline_run}
\alias{az_pipeline_run}
\title{Azure DevOps Pipeline Run Class}
\usage{
az_pipeline_run(
  id = integer(0),
  pipeline_id = 0L,
  name = "",
  status = "inProgress",
  result = "",
  created_date = "",
  web_url = "",
  logs_url = ""
)
}
\description{
Represents an execution run of an Azure Pipeline.
}
