% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{az_pull_request_get}
\alias{az_pull_request_get}
\title{Get a Single Pull Request}
\usage{
az_pull_request_get(
  pull_request_id,
  repository_id = NULL,
  project = NULL,
  client = NULL
)
}
\arguments{
\item{pull_request_id}{Integer ID of the pull request.}

\item{repository_id}{Optional repository name or ID.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
An S7 \code{az_pull_request} object.
}
\description{
Get a Single Pull Request
}
