% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{az_repo}
\alias{az_repo}
\title{Azure DevOps Repository Class}
\usage{
az_repo(
  id = character(0),
  name = character(0),
  default_branch = "main",
  web_url = "",
  project = ""
)
}
\description{
Represents a Git repository in Azure Repos.
}
