% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{az_repo_get}
\alias{az_repo_get}
\title{Get a Single Git Repository}
\usage{
az_repo_get(
  repository_id,
  project = NULL,
  client = NULL,
  call = rlang::caller_env()
)
}
\arguments{
\item{repository_id}{Repository name or ID.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}

\item{call}{Caller environment for error attribution.}
}
\value{
An S7 \code{az_repo} object.
}
\description{
Get a Single Git Repository
}
