% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{az_test_plan}
\alias{az_test_plan}
\title{Azure DevOps Test Plan Class}
\usage{
az_test_plan(
  id = integer(0),
  name = character(0),
  state = "Active",
  area_path = "",
  iteration = "",
  web_url = ""
)
}
\description{
Represents a test plan in Azure Test Plans.
}
