% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{az_test_run}
\alias{az_test_run}
\title{Azure DevOps Test Run Class}
\usage{
az_test_run(
  id = integer(0),
  name = character(0),
  state = "InProgress",
  total_tests = 0L,
  pass_rate = 0,
  web_url = ""
)
}
\description{
Represents an executed test run in Azure Test Plans.
}
