% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_plans.R
\name{az_test_run_metrics_get}
\alias{az_test_run_metrics_get}
\title{Get Test Run Metrics & Summary}
\usage{
az_test_run_metrics_get(run_id, project = NULL, client = NULL)
}
\arguments{
\item{run_id}{Integer ID of the test run.}

\item{project}{Project name or ID.}

\item{client}{Optional \code{az_client} S7 object.}
}
\value{
An S7 \code{az_test_run} object.
}
\description{
Get Test Run Metrics & Summary
}
