% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{az_work_item}
\alias{az_work_item}
\title{Azure DevOps Work Item Class}
\usage{
az_work_item(
  id = integer(0),
  rev = 1L,
  type = "",
  title = "",
  state = "",
  assigned_to = "",
  url = "",
  web_url = "",
  fields = list()
)
}
\description{
Represents a work item (Bug, Task, User Story, Epic, Feature) in Azure Boards.
}
